+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- metrics registry refactoring to search with `O(1)`
+
 ### Fixed
 - be gentle to http routes, don't leave gaps in the array
   [#246](https://github.com/tarantool/metrics/issues/246)