+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - be gentle to http routes, don't leave gaps in the array
   [#246](https://github.com/tarantool/metrics/issues/246)
 
+### Added
+- `tnt_clock_delta` metric to compute clock difference on instances
+
 ## [0.9.0] - 2021-05-28
 ### Fixed
 - cartridge metrics role fails to start without http [#225](https://github.com/tarantool/metrics/issues/225)